+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2024-03-05
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `no-image` for result meal without image URL.
+- [Marek Kurańda](https://github.com/mjkuranda): Meal results at search page.
+
 ## [1.7.0] - 2024-03-04
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Custom `not-found` page.